requirements.txt
click==6.7
pandas==0.20.3
Wand==0.4.4
tensorflow_gpu==1.3.0
gensim==3.0.1
nltk==3.2.5
pytesseract==0.1.7
numpy==1.13.1
Image==1.5.16
Pillow==4.3.0
python-dotenv==0.7.1
scikit_learn==0.19.1
tensorflow==1.4.0
